optimization applied

+++ params.h
@@ -1,7 +1,7 @@
 #pragma once 
-
+#include "hls_half.h"
 #define Capa_In_Size 4
 #define Infrared_In_Size 16
 #define Output_Size 2
 #define DataSize 393
-typedef double DataType;
+typedef float DataType;
